+++ Card/Card.js
@@ -29,6 +29,7 @@ import Image from '../Image';
 import $L from '../internal/$L';
 import AsyncRenderChildren from '../internal/AsyncRenderChildren';
 import {Marquee, MarqueeController} from '../Marquee';
+import ProgressBar from '../ProgressBar';
 import Skinnable from '../Skinnable';
 
 import componentCss from './Card.module.less';
@@ -213,6 +214,14 @@ const CardBase = kind({
 		primaryBadgeSrc: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
 
 		/**
+		 * The progress displayed inside the ProgressBar
+		 *
+		 * @type {Number}
+		 * @public
+		 */
+		progress: PropTypes.number,
+
+		/**
 		 * Set to `true` to display the image with rounded corners.
 		 *
 		 * @type {Boolean}
@@ -242,7 +251,15 @@ const CardBase = kind({
 		 * @type {Boolean}
 		 * @public
 		 */
-		selected: PropTypes.bool
+		selected: PropTypes.bool,
+
+		/**
+		 * Activates the 'ProgressBar'.
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		showProgressBar: PropTypes.bool
 	},
 	defaultProps: {
 		icon: 'check',
@@ -270,7 +287,7 @@ const CardBase = kind({
 			return `${children || ''}${label ? ` ${label}` : ''}${secondaryLabel ? ` ${secondaryLabel}` : ''}${selected ? ' ' + $L('Selected') : ''}`;
 		},
 		captionOverlay: ({captionOverlay, captionOverlayOnFocus}) => captionOverlay || captionOverlayOnFocus,
-		children: ({centered, children, css, 'data-index': index, imageIconSrc, label, orientation, secondaryLabel}) => {
+		children: ({centered, children, css, 'data-index': index, imageIconSrc, label, orientation, progress, secondaryLabel, showProgressBar}) => {
 			const hasImageIcon = imageIconSrc && orientation === 'vertical';
 			const alignment = centered && !imageIconSrc ? {alignment: 'center'} : null;
 
@@ -288,6 +305,7 @@ const CardBase = kind({
 						<Marquee {...alignment} className={css.caption} marqueeOn="hover">{children}</Marquee>
 						{typeof label !== 'undefined' ? <Marquee {...alignment} className={css.label} marqueeOn="hover">{label}</Marquee> : null}
 						{typeof secondaryLabel !== 'undefined' ? <Marquee {...alignment} className={css.label} marqueeOn="hover">{secondaryLabel}</Marquee> : null}
+						{showProgressBar ? <ProgressBar progress={progress} /> : null}
 					</Cell>
 				</Row>
 			);
@@ -324,7 +342,9 @@ const CardBase = kind({
 		delete rest.captionOverlayOnFocus;
 		delete rest.centered;
 		delete rest.label;
+		delete rest.progress;
 		delete rest.secondaryLabel;
+		delete rest.showProgressBar;
 		delete rest.imageIconSrc;
 		delete rest.hasContainer;
 		delete rest.roundedImage;