+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 ### Added
 
 - `limestone/Card` `icon` prop to allow user to change the icon
+- `limestone/Card` added option for progressBar
 
 ## [1.6.0] - 2025-10-14